requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.10.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
blessed==1.20.0
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
googlesearch-python==1.2.3
idna==3.4
inquirer==3.1.3
magic-filter==1.0.9
multidict==6.0.4
python-editor==1.0.4
python-googlesearch==1.1.1
pytz==2023.3
pyuseragents==1.0.5
readchar==4.0.5
redis==4.5.5
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
tzlocal==5.0.1
urllib3==2.0.2
wcwidth==0.2.6
yarl==1.9.2
